Moved Wrapper logic to Container (issue #171)
