+++ src/sass/Offers.scss
@@ -129,7 +129,7 @@ $card-head-dark: rgba(0, 0, 0, 0.65);
   }
 }
 
-.load {
+.loading {
   .head {
     animation-duration: 1s;
     animation-fill-mode: forwards;
