+++ src/ts/Offers/OffersList.tsx
@@ -6,15 +6,15 @@ import Offer from './Offer';
 export default function OffersList(props) {
   const {
     list,
-    load,
     date,
+    loading,
     location,
   } = props;
 
   const items = list.map((item, key) => <Offer data={item} key={key} date={date} location={location}/>);
 
   return (
-    <div className={`offers ${load ? 'load' : ''}`}>
+    <div className={`offers ${loading ? 'loading' : ''}`}>
       {items}
     </div>
   );
